requirements.txt
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
concurrent-log-handler==0.9.28
ddddocr==1.5.6
dramatiq==1.18.0
dynaconf==3.2.11
fake-http-header==0.3.5
filelock==3.19.1
flatbuffers==25.2.10
gevent==25.5.1
greenlet==3.2.4
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
linktools==0.8.17
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
packaging==25.0
pika==1.3.2
pillow==11.3.0
pip-tools==7.5.0
playwright==1.54.0
pluggy==1.6.0
portalocker==3.2.0
prometheus_client==0.22.1
protobuf==6.32.0
psutil==7.0.0
pyee==13.0.0
Pygments==2.19.2
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dotenv==1.1.1
python-slugify==8.0.4
redis==6.4.0
regex==2025.7.34
requests==2.32.4
rich==14.1.0
setuptools==80.9.0
sympy==1.14.0
text-unidecode==1.3
tf-playwright-stealth==1.2.0
typing_extensions==4.14.1
urllib3==2.5.0
watchdog==6.0.0
watchdog_gevent==0.2.1
wheel==0.45.1
zope.event==5.1.1
zope.interface==7.2
